refactor(common components) : file structure
