+++ src/app/components/common/NewRecipe/RecipeSteps/RecipeStepsContent.module.scss
@@ -1,0 +1,18 @@
+@use '@/styles/theme.scss' as *;
+
+.steps {
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+  gap: 10px;
+}
+
+.step {
+  flex: 1;
+  display: flex;
+  gap: 5px;
+
+  &__value {
+    flex-grow: 1;
+  }
+}
