+++ src/app/components/common/NewRecipe/SearchProductModal/ExistingProduct.tsx
