+++ src/app/components/common/NewRecipe/SearchProductModal/NewProduct.tsx
