+++ src/app/components/common/NewRecipe/SearchProductModal/style.module.scss
