+++ src/app/components/common/NewRecipe/helper.ts
@@ -1,5 +1,14 @@
-import { ImgState, NewRecipeIngredientStates, TextInputs } from '.';
-import { Step } from './components/RecipeStepsContent';
+import {
+  ImgState,
+  NewRecipeIngredientState,
+  NewRecipeIngredientStates,
+  TextInputs,
+} from '.';
+import {
+  SelectedIngredientInfo,
+  SelectedProductInfo,
+} from './RecipeIngredients/RecipeIngredientsContent';
+import { Step } from './RecipeSteps/RecipeStepsContent';
 
 export const getValidIngredients = (ingredients: NewRecipeIngredientStates) =>
   ingredients.filter(
@@ -30,3 +39,54 @@ export const checkIfAllFieldsAreFilled = ({
   img &&
   ingredients[0].ingredientName.trim() &&
   steps[0].value.trim().length;
+
+export const removeProductInfoFromSelectedIngredient = (
+  ingredients: NewRecipeIngredientStates,
+  selectedIngredientId: NewRecipeIngredientState['id'],
+): NewRecipeIngredientStates =>
+  ingredients.map((item) =>
+    item.id === selectedIngredientId
+      ? {
+          ...item,
+          newProduct: null,
+          productId: null,
+        }
+      : item,
+  );
+
+interface AddProductProps {
+  selectedIngredientInfo: SelectedIngredientInfo;
+  selectedIngredient: NewRecipeIngredientState;
+  selectedProductInfo: SelectedProductInfo;
+  ingredients: NewRecipeIngredientStates;
+}
+
+export const addProductInfoToSelectedIngredient = ({
+  ingredients,
+  selectedIngredient,
+  selectedProductInfo,
+}: AddProductProps): NewRecipeIngredientStates =>
+  ingredients.map((item) =>
+    item.id === selectedIngredient.id
+      ? {
+          ...item,
+          ingredientName: selectedProductInfo?.ingredientName ?? '',
+          ingredientId: selectedProductInfo?.ingredientId ?? null,
+          productId: selectedProductInfo?.productId ?? null,
+          newProduct: selectedProductInfo?.newProduct ?? null,
+        }
+      : item,
+  );
+
+export const onFieldChange = <T extends { id: string }>(
+  setState: React.Dispatch<React.SetStateAction<T[]>>,
+  id: string,
+  fieldName: string,
+  value: string,
+) => {
+  setState((prev) =>
+    prev.map((item) =>
+      item.id === id ? { ...item, [fieldName]: value } : item,
+    ),
+  );
+};
